refactor(e2e): split helpers.ts into focused modules
Split delete-rule helpers into logical modules for better maintainability:
- constants.ts: Delete-specific timeout constants
- dialog-operations.ts: Confirm dialog interaction helpers
- rule-operations.ts: Rule table query and wait helpers
- helpers.ts: Re-exports for backward compatibility
