+++ host-frontend-root/frontend-src-root/tests/e2e/pages/rule-list/features/delete-rule/constants.ts
@@ -1,0 +1,12 @@
+/**
+ * 削除機能固有の定数
+ */
+
+/** 確認ダイアログ表示待機のタイムアウト（ms） */
+export const CONFIRM_DIALOG_TIMEOUT = 10000;
+
+/** 削除完了待機のタイムアウト（ms） */
+export const DELETE_COMPLETE_TIMEOUT = 10000;
+
+/** 要素可視性待機のタイムアウト（ms）- toPass内での使用を想定 */
+export const ELEMENT_VISIBILITY_TIMEOUT = 5000;
